[incubator/jaeger] Adds ability to utilize a cloud providers firewall for services (#3457)
* Adds ability to utilize a cloud providers firewall for services

* Adds declaration allowing the user to configure which IP CIDRs are
allowed to access a service via a cloud providers firewall
* An example configuration of `values.yaml` may look like this:
```
query:
  service:
    type: LoadBalancer
    loadBalancerSourceRanges:
      - "192.0.2.0/24"
      - "198.51.100.0/24"
```
* The above would configure a cloud provider's firewall with a ruleset
allowing only the configured IP ranges access to the query service
* This adds this declaration across the board for all services as it may
be valuable. One example, is to send metric data from applications that
do not reside in kubernetes
* This PR also updates the README to reflect other options for services
that have not been updated over time.

* rids empty lines

* Utilizes a template for the lb source ranges, passes in appropriate
scope to allow each service to be configured individually if required by
administrator

* missed an empty line